Split types out from schema file, prep for further refactors
